chore: 테스트용 섹션을 TestSection 컴포넌트로 분리 	modified:   src/App.tsx 	new file:   src/components/TestSection.tsx
